test : 변경 사항 기반 단위 테스트 수정 #7
